+++ src/main/kotlin/syntax/backend/runways/service/FineDustService.kt
@@ -3,5 +3,5 @@ package syntax.backend.runways.service
 import syntax.backend.runways.dto.FineDustDataDTO
 
 interface FineDustService {
-    fun getFineDustData(x:Int, y:Int): FineDustDataDTO
+    fun getFineDustData(x:Double, y:Double): FineDustDataDTO
 }
